requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
pilkit==2.0
Pillow==6.0.0
Django==2.1.7
django-allauth==0.39.1
django-appconf==1.0.3
django-imagekit==4.0.2
django-markdownx==2.0.28
django-mdeditor==0.1.14
django-widget-tweaks==1.4.3
django-rest-framework==3.9.4
idna==2.8
imagekit==1.0.6
Markdown==3.1
markdown2==2.3.7
oauthlib==3.0.1
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.24.3
wfastcgi==3.0.0
